+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-07-17T06:33:15.314Z")
+    created: Optional[datetime] = parse_datetime("2025-07-21T05:52:38.303Z")
     r"""The date and time the resource was created"""