+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K0BHCS1EX13HQTKD4HJBRMJH"
+        "01K0NRN9J3RW9NPNZQ706RF5CH"
     )
     r"""The id of the resource"""