+++ packages/orq-rc/RELEASES.md
@@ -2019,3 +2019,13 @@ Based on:
 - [python v3.10.0-rc.10] packages/orq-rc
 ### Releases
 - [PyPI v3.10.0-rc.10] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.10 - packages/orq-rc
+
+## 2025-07-21 05:54:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.583.0 (2.658.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.10.0-rc.11] packages/orq-rc
+### Releases
+- [PyPI v3.10.0-rc.11] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.11 - packages/orq-rc